Remove custom logger
Though nice looking, I do not think that `request-promise-retry` should
be configuring a custom logging.
Possibly, it could still keep a prefix for its logs, but IMO it should
not define specific rules.

The application using this nice library is the place to customize logging.

First time I realized this and was surprising is when I had bumped the min
log level to say warn, and was still seeing the retry related logs coming
from here.